+++ src/synapsis.cpp
@@ -22,7 +22,7 @@ using namespace neuralpp;
  * @param a Activation function
  * @param d Derivate for activation function
  */
-Synapsis::Synapsis (Neuron* i, Neuron* o, float(*a)(float), float(*d)(float))  {
+Synapsis::Synapsis (Neuron* i, Neuron* o, double(*a)(double), double(*d)(double))  {
 	srand((unsigned) time(NULL));
 
 	delta=0;
@@ -42,7 +42,7 @@ Synapsis::Synapsis (Neuron* i, Neuron* o, float(*a)(float), float(*d)(float))  {
  * @param a Activation function
  * @param d Derivate for activation function
  */
-Synapsis::Synapsis (Neuron* i, Neuron* o, float w, float(*a)(float), float(*d)(float))  {
+Synapsis::Synapsis (Neuron* i, Neuron* o, double w, double(*a)(double), double(*d)(double))  {
 	delta=0;
 	weight=w;
 	in=i;
@@ -65,21 +65,21 @@ Neuron* Synapsis::getOut()  { return out; }
 /**
  * @return Weight of the synapsis
  */
-float Synapsis::getWeight()  { return weight; }
+double Synapsis::getWeight()  { return weight; }
 
 /**
  * @return Delta of the synapsis
  */
-float Synapsis::getDelta()  { return delta; }
+double Synapsis::getDelta()  { return delta; }
 
 /**
  * @brief It sets the weight of the synapsis
  */
-void Synapsis::setWeight(float w)  { weight=w; }
+void Synapsis::setWeight(double w)  { weight=w; }
 
 /**
  * @brief It sets the delta (how much to change the weight after an update) 
  *  of the synapsis
  */
-void Synapsis::setDelta(float d)  { delta=d; }
+void Synapsis::setDelta(double d)  { delta=d; }
